Add horizontal scrollable feature
