+++ app/src/App.css
@@ -1,3 +1,4 @@
 * {
   margin: 0;
+  /* background-color: red; */
 }
